backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pillow==11.3.0
PyMuPDF==1.26.4
pypdf==5.9.0
reportlab==4.4.3
pydantic-settings==2.10.1
